chore: Complete Phase 2 (Mechanical Conversion) of Python 2→3 migration
Mechanically convert all 44 modules across 9 conversion units from
Python 2 to Python 3.12 syntax. Gate check: PASS 7/7 criteria.

699 Py2-isms eliminated from executable code. Key transformations:
- print statements→print(), except comma→as, raise 2-arg→constructor
- dict.has_key→in, iteritems/iterkeys/itervalues→items/keys/values
- xrange→range, raw_input→input, long→int, unicode→str
- 19 stdlib import renames (ConfigParser, cPickle, imp, Queue, etc.)
- 22 implicit relative imports→explicit in 6 __init__.py files

Critical data-layer fixes:
- database.py: str(row[N])→bytes(row[N]) for SQLite BLOB deserialization
- csv_processor.py: binary mode→text mode CSV rewrite for Py3
- types.py: __cmp__→__eq__/__lt__, __nonzero__→__bool__, metaclass= syntax
- plugin_loader.py: imp.load_source()→importlib spec_from_file_location
- compat.py: simplified to Py3-only (removed all try/except branches)
- string_helpers.py: complete bytes/str/unicode type-check redesign

Files parse as valid Python 3.12 but have not been executed yet —
runtime validation is Phase 3. from __future__ imports retained
intentionally for Phase 4 removal.

+++ tests/test_core_types_characterization.py
@@ -151,20 +151,20 @@ class TestLargeCounter:
     @pytest.mark.py2_behavior
     def test_construction_with_long(self):
         """Captures: initial value stored as long(). long removed in Py3."""
-        counter = LargeCounter(0L)
+        counter = LargeCounter(0)
         assert counter.value == 0
 
     @pytest.mark.py2_behavior
     def test_increment_with_long_literal(self):
         """Captures: increment(1L) uses long literals."""
         counter = LargeCounter()
-        counter.increment(1L)
+        counter.increment(1)
         assert counter.value == 1
 
     @pytest.mark.py2_behavior
     def test_max_value_long_literal(self):
         """Captures: MAX_VALUE is 2**64-1 as long literal."""
-        assert LargeCounter.MAX_VALUE == 18446744073709551615L
+        assert LargeCounter.MAX_VALUE == 18446744073709551615
 
     def test_wraps_on_overflow(self):
         """Captures: counter wraps at MAX_VALUE + 1."""
@@ -176,7 +176,7 @@ class TestLargeCounter:
     def test_long_method(self):
         """Captures: __long__ returns counter value. Removed in Py3."""
         counter = LargeCounter(42)
-        assert long(counter) == 42
+        assert int(counter) == 42
 
     def test_type_check_on_init(self):
         """Captures: non-integer raises TypeError."""
